fix proximity
